pyproject.toml
[tool.poetry]
name = "ansible-github"
version = "0.0.1"
description = "Ansible collection for working with Github."
authors = ["jheddings <jheddings@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "plugins/modules"}]
repository = "https://github.com/jheddings/ansible-github"

[tool.poetry.dependencies]
python = "^3.9"
ansible = "^7.2.0"
pygithub = "^1.57"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.1.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
